requirements.txt
dotenv
requests
Pillow
pandas
